+++ Modules/Intent.Modules.AspNetCore.Controllers/release-notes.md
@@ -1,3 +1,8 @@
+### Version 5.2.0
+
+- FileBuilder metadata updated.
+- Updated contract to include Api Version information.
+
 ### Version 5.1.1
 
 - Have controllers fully manage using statements by default.