+++ Modules/Intent.Modules.AspNetCore.Versioning/release-notes.md
@@ -1,1 +1,3 @@
 ### Version 1.0.0
+
+- New Module to manage API versions for Services, Commands and Queries.